dpp-backend/scripts/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Flask==3.0.2
identify==2.5.35
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
platformdirs==4.2.0
PyYAML==6.0.1
requests==2.31.0
urllib3==2.2.0
Werkzeug==3.0.1
